pyproject.toml
[project]
name = "dota-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Matvey",email = "mateusru54063@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)"
]

[tool.poetry]
packages = [{include = "dota_project", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
